+++ backend/app/routes/user.py
@@ -1,0 +1,8 @@
+from fastapi import APIRouter
+
+router = APIRouter()
+
+
+@router.get("/")
+async def read_root() -> dict:
+    return {"message": "User endpoint"}
