+++ backend/app/schemas/article.py
