+++ backend/app/schemas/user.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel
+
+
+class User(BaseModel):
+    id: int
+    username: str
+    email: str
+    hashed_password: str
